[FIX] Allgather cont_send.

+++ bench/run_allgather.sh
@@ -84,16 +84,17 @@ do
     
     # ATTENTION: Showing decision process adds non-negligible overhead (for small vectors). Use it with care.
     # TODO: Maybe I should prealloc only for large ALLGATHER?
-    EXTRA_MCAS="" #"-mca mpi_print_stats 1 -mca coll_select_show_decision_process 2" #"-mca coll_base_reduce_commute_safe 1"
+    # Also if it is for bcast, according to the documentation should be helpful even for allgather if all the counts are equal
+    EXTRA_MCAS="-mca coll_tuned_bcast_same_count=1" #"-mca mpi_print_stats 1 -mca coll_select_show_decision_process 2" #"-mca coll_base_reduce_commute_safe 1"
 
-    # Do a run just to print the decision process
+    ## Do a run just to print the decision process
     DEFAULT_ALGO="default"
     LIBSWING_ALLGATHER_ALGO_FAMILY="DEFAULT" ${MPIRUN} -mca coll_select_show_decision_process 2 -mca coll_tuned_prealloc_size ${coll_tuned_prealloc_size} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} ${DATATYPE} ${actual_count} 1
     ALGO_FNAME=default-${DEFAULT_ALGO}
     mv ${OUT_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.decision; rm -f ${OUT_PREFIX}* 
     if [ -f ${ERR_PREFIX}*.0 ]; then mv ${ERR_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.decision.err; rm -f ${ERR_PREFIX}*; fi
 
-    # Run the actual benchmark
+    ## Run the actual benchmark
     DEFAULT_ALGO="default"
     LIBSWING_ALLGATHER_ALGO_FAMILY="DEFAULT" ${MPIRUN} ${EXTRA_MCAS} -mca coll_tuned_prealloc_size ${coll_tuned_prealloc_size} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} ${DATATYPE} ${actual_count} ${iterations}
     ALGO_FNAME=default-${DEFAULT_ALGO}
@@ -131,20 +132,36 @@ do
                     if [ -f ${ERR_PREFIX}*.0 ]; then mv ${ERR_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.err; rm -f ${ERR_PREFIX}*; fi
                 fi
             done
+
+            # Run swing CONT_SEND (only works for 1 port)
+            if [ $PORTS -eq 1 ]; then
+                export LIBSWING_ALLGATHER_ALGO_FAMILY="SWING" 
+                export LIBSWING_ALLGATHER_ALGO_LAYER="UTOFU" 
+                export LIBSWING_ALLGATHER_ALGO="VEC_DOUBLING_CONT_SEND"
+                for SEGMENT_SIZE in 0 #4096 65536 1048576
+                do                
+                    if [ $SEGMENT_SIZE -lt $total_msg_size ]; then
+                        LIBSWING_SEGMENT_SIZE=${SEGMENT_SIZE} ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} ${DATATYPE} ${actual_count} ${iterations}                    
+                        ALGO_FNAME=${LIBSWING_ALLGATHER_ALGO_FAMILY}-${LIBSWING_ALLGATHER_ALGO}-${LIBSWING_ALLGATHER_ALGO_LAYER}-${SEGMENT_SIZE}-${PORTS}
+                        mv ${OUT_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.csv; rm -f ${OUT_PREFIX}* 
+                        if [ -f ${ERR_PREFIX}*.0 ]; then mv ${ERR_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.err; rm -f ${ERR_PREFIX}*; fi
+                    fi
+                done          
+            fi
         
-            # Run recdoub
-            export LIBSWING_ALLGATHER_ALGO_FAMILY="RECDOUB" 
-            export LIBSWING_ALLGATHER_ALGO_LAYER="UTOFU" 
-            export LIBSWING_ALLGATHER_ALGO="VEC_DOUBLING_CONT_PERMUTE"    
-            for SEGMENT_SIZE in 0 #4096 65536 1048576
-            do                
-                if [ $SEGMENT_SIZE -lt $total_msg_size ]; then
-                    LIBSWING_SEGMENT_SIZE=${SEGMENT_SIZE} ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} ${DATATYPE} ${actual_count} ${iterations}                    
-                    ALGO_FNAME=${LIBSWING_ALLGATHER_ALGO_FAMILY}-${LIBSWING_ALLGATHER_ALGO}-${LIBSWING_ALLGATHER_ALGO_LAYER}-${SEGMENT_SIZE}-${PORTS}
-                    mv ${OUT_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.csv; rm -f ${OUT_PREFIX}* 
-                    if [ -f ${ERR_PREFIX}*.0 ]; then mv ${ERR_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.err; rm -f ${ERR_PREFIX}*; fi
-                fi
-            done
+            ## Run recdoub
+            #export LIBSWING_ALLGATHER_ALGO_FAMILY="RECDOUB" 
+            #export LIBSWING_ALLGATHER_ALGO_LAYER="UTOFU" 
+            #export LIBSWING_ALLGATHER_ALGO="VEC_DOUBLING_CONT_PERMUTE"    
+            #for SEGMENT_SIZE in 0 #4096 65536 1048576
+            #do                
+            #    if [ $SEGMENT_SIZE -lt $total_msg_size ]; then
+            #        LIBSWING_SEGMENT_SIZE=${SEGMENT_SIZE} ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} ${DATATYPE} ${actual_count} ${iterations}                    
+            #        ALGO_FNAME=${LIBSWING_ALLGATHER_ALGO_FAMILY}-${LIBSWING_ALLGATHER_ALGO}-${LIBSWING_ALLGATHER_ALGO_LAYER}-${SEGMENT_SIZE}-${PORTS}
+            #        mv ${OUT_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.csv; rm -f ${OUT_PREFIX}* 
+            #        if [ -f ${ERR_PREFIX}*.0 ]; then mv ${ERR_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.err; rm -f ${ERR_PREFIX}*; fi
+            #    fi
+            #done
         fi
     done
     echo " ${GREEN}[Done]${NC}"
